Started Algo Implementation
Added Form Name
Added Algos to drop down
Added Algos Folder w/ Kruskal code
Getting inaccessible errors -.-

+++ treeCalc/Path.cs
@@ -8,16 +8,25 @@ namespace treeCalc
 {
     class Path
     {
+
+        #region Public Properties
+
         public Node StartNode { get; set; }
         public Node EndNode { get; set; }
-        public double Weight { get; set; }
+        public int Weight { get; set; }
 
-        public Path(Node startNode, Node endNode, double weight)
+        #endregion
+
+        #region Constructor
+
+        public Path(Node startNode, Node endNode, int weight)
         {
             this.StartNode = startNode;
             this.EndNode = endNode;
             this.Weight = weight;
         }
+
+        #endregion
 
         public override string ToString()
         {
